pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-multi-agent"
version = "0.1.0"
authors = [
  { name="Flow", email="flow@example.com" }
]
description = "Sistema Multi-Agent usando Google Gemini"
readme = "README.md"
requires-python = ">=3.9,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-generativeai",
    "python-dotenv",
    "rich",
    "agentops==0.3.20"
]

[tool.setuptools]
packages = ["gemini"]
